define evil properties on my custom navigation functions
Evil listens for interactively called functions by adding a hook to the
'pre-command-hook, and checking each of those functions for evil attributes. If
you wrap an evil function with a custom function, their hook will not see it, so
you need to apply the same attributes to your wrapper function.

In this case the specific bug I was working around was pressing n to continue to
the next search item would override the `.` repeat last command item, so you
couldn't quickly cycle through items in a search with `n` and then `.` to repeat
the same change to each of them (instead with this bug the `.` would repeat the
find next item rather than repeating making the change).